Redundant file cleanup Part 8
