--- .dockerignore
@@ -1,62 +1,0 @@
-# Version control
-.git
-.gitignore
-
-# Python
-__pycache__/
-*.py[cod]
-*$py.class
-*.so
-.Python
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
-.eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-wheels/
-*.egg-info/
-.installed.cfg
-*.egg
-
-# Virtual environments
-venv/
-env/
-ENV/
-simple-venv/
-
-# Testing
-.pytest_cache/
-.coverage
-htmlcov/
-.tox/
-
-# Docker
-.dockerignore
-Dockerfile
-docker-compose.yml
-
-# Logs
-*.log
-logs/
-
-# Local development
-.env
-.env.local
-.DS_Store
-
-# Test files and results
-test_files/
-test_results/
-plots/
-
-# IDE
-.idea/
-.vscode/
-*.swp
-*.swo
