+++ .gitignore
@@ -13,21 +13,10 @@ __pycache__/**
 
 # Application specific
 *.log
-test_results/
-test_files/uploaded/
 .env
 *.env
 *.env.*
 
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*.cover
 
 # OS
 .DS_Store?
