--- Dockerfile
@@ -1,29 +1,0 @@
-FROM python:3.11.11
-
-WORKDIR /app
-
-# Install system dependencies
-RUN apt-get update && apt-get install -y --no-install-recommends \
-    build-essential \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy requirements first for better caching
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy application code
-COPY . .
-
-# Create directory for temporary uploads
-RUN mkdir -p /tmp/uploads && chmod 777 /tmp/uploads
-
-# Set environment variables
-ENV FLASK_APP=wsgi.py
-ENV FLASK_ENV=production
-ENV PORT=8080
-
-# Expose port
-EXPOSE 8080
-
-# Run with gunicorn
-CMD ["gunicorn", "wsgi:app", "--bind", "0.0.0.0:8080", "--log-level", "info"]
