+++ requirements.txt
@@ -19,6 +19,8 @@ PyPDF2==3.0.1
 docx2txt==0.8
 pdfminer.six==20221105
 
+pkg_resources
+
 # Security & Networking
 urllib3==2.0.7
